Exec: Rework kill command
Apparently the process structure of MSYS2 has changed and the first bash
process is no longer available to kill. Killing it from Windows does not
work because all children will keep running.

We now look through the children of the bash process (if any) and
terminate these from within MSYS2.